feat: 재사용을 위한 SplittedClockView 의 부모 뷰 ClockView 정의
